Move ISS coordinator to separate module
- Add `coordinator.py` with `IssCoordinator` (typed `config_entry`
  at class level), `IssData`
- Update `sensor.py` to import directly from `.coordinator`,
  avoiding re-export through `__init__`
