+++ homeassistant/components/iss/__init__.py
@@ -2,63 +2,21 @@
 
 from __future__ import annotations
 
-from dataclasses import dataclass
-from datetime import timedelta
-import logging
-
-import pyiss
-import requests
-from requests.exceptions import HTTPError
-
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import Platform
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
 from .const import DOMAIN
-
-_LOGGER = logging.getLogger(__name__)
+from .coordinator import IssCoordinator
 
 PLATFORMS = [Platform.SENSOR]
-
-
-@dataclass
-class IssData:
-    """Dataclass representation of data returned from pyiss."""
-
-    number_of_people_in_space: int
-    current_location: dict[str, str]
-
-
-def update(iss: pyiss.ISS) -> IssData:
-    """Retrieve data from the pyiss API."""
-    return IssData(
-        number_of_people_in_space=iss.number_of_people_in_space(),
-        current_location=iss.current_location(),
-    )
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Set up this integration using UI."""
     hass.data.setdefault(DOMAIN, {})
 
-    iss = pyiss.ISS()
-
-    async def async_update() -> IssData:
-        try:
-            return await hass.async_add_executor_job(update, iss)
-        except (HTTPError, requests.exceptions.ConnectionError) as ex:
-            raise UpdateFailed("Unable to retrieve data") from ex
-
-    coordinator = DataUpdateCoordinator(
-        hass,
-        _LOGGER,
-        config_entry=entry,
-        name=DOMAIN,
-        update_method=async_update,
-        update_interval=timedelta(seconds=60),
-    )
-
+    coordinator = IssCoordinator(hass, entry)
     await coordinator.async_config_entry_first_refresh()
 
     hass.data[DOMAIN] = coordinator
