+++ homeassistant/components/iss/sensor.py
@@ -11,13 +11,10 @@ from homeassistant.const import ATTR_LATITUDE, ATTR_LONGITUDE, CONF_SHOW_ON_MAP
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.device_registry import DeviceEntryType, DeviceInfo
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
-from homeassistant.helpers.update_coordinator import (
-    CoordinatorEntity,
-    DataUpdateCoordinator,
-)
+from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from . import IssData
 from .const import DEFAULT_NAME, DOMAIN
+from .coordinator import IssCoordinator
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -28,14 +25,14 @@ async def async_setup_entry(
     async_add_entities: AddConfigEntryEntitiesCallback,
 ) -> None:
     """Set up the sensor platform."""
-    coordinator: DataUpdateCoordinator[IssData] = hass.data[DOMAIN]
+    coordinator: IssCoordinator = hass.data[DOMAIN]
 
     show_on_map = entry.options.get(CONF_SHOW_ON_MAP, False)
 
     async_add_entities([IssSensor(coordinator, entry, show_on_map)])
 
 
-class IssSensor(CoordinatorEntity[DataUpdateCoordinator[IssData]], SensorEntity):
+class IssSensor(CoordinatorEntity[IssCoordinator], SensorEntity):
     """Implementation of the ISS sensor."""
 
     _attr_has_entity_name = True
@@ -43,7 +40,7 @@ class IssSensor(CoordinatorEntity[DataUpdateCoordinator[IssData]], SensorEntity)
 
     def __init__(
         self,
-        coordinator: DataUpdateCoordinator[IssData],
+        coordinator: IssCoordinator,
         entry: ConfigEntry,
         show: bool,
     ) -> None:
